py/projects/acuity_model_server/pyproject.toml
[tool.poetry]
name = "acuity_model_server"
version = "0.1.0"
description = ""
authors = ["Manu Sabherwal <manu.sabherwal@*company-data-covered*.com>"]

[tool.poetry.dependencies]
python = ">=3.10.12 <3.11"
proto = {path = "../../lib/proto", develop = true}
auth = {path = "../../lib/auth", develop = true}
acuity_model = {path = "../../lib/acuity_model", develop = true}
monitoring = {path = "../../lib/monitoring", develop = true}
python-decouple = "^3.6"
python-json-logger = "^2.0.4"
ddtrace = "^1.7.3"
grpcio-health-checking = "1.47.2"
grpcio-testing = "1.47.2"
statsig = "^0.16.11"
# requests is imported by dependencies. we force the latest version for security
requests = "^2.31.0"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"
ipdb = "^0.13.9"
pytest-asyncio = "^0.18.3"

[tool.poetry.group.dev.dependencies]
pytest-cov = "^4.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = "strict"
